fix: matching of octal escape seqs in regex
...for detects the IE octal, hex & unicode entities.

The change made is to include optional '0' characters
before the hexadecimal value in the '\u' escape sequence,
and also to include matching for octal escape sequences
(e.g '\123') along with '\x' and '\u' sequences.

This will improve handling of unicode escape codes in the
application and make it more robust.