feat: add rule for common event attrs & props (ID: 80)
This commit introduces a detection mechanism for common event
attributes and properties with ID 80. The rule has been
expanded to encompass a range of events and attributes. The
impact has been rated as 6, denoting the potential significance
of this detection.

I would like to express my gratitude to Aneesh Nadh (@ma1f0y)
for responsibly reporting a detection bypass. Their efforts in
identifying and bringing attention to this issue are greatly
appreciated.

Thanks to their help, we were able to promptly address the
bypass and improve the overall security and reliability of
our rule.

Signed-off-by: Dwi Siswanto <me@dw1.io>